requirements.txt
# Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Config
pyyaml>=6.0

# Data & CLI
python-dotenv>=1.0.0
